* move `ItemFilter` to it's own class to make it more flexible * some more simplifications * add types to the ItemFilter * add MIGRATION note's
